backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
pymongo==4.6.0
Pillow==10.1.0
pytesseract==0.3.10
openai==1.3.7
langchain==0.0.350